pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rouvy-api-client"
version = "0.1.0"
description = "A Python client for the Rouvy API with support for authentication, session management, and parsing of turbo-stream formatted responses"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Mike Hill" }]
dependencies = ["requests>=2.31.0", "python-dotenv>=1.0.0"]

[project.scripts]
rouvy-api = "rouvy_api_client.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "responses>=0.25.0",
    "aiohttp>=3.9.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "types-requests>=2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rouvy_api_client"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/test_cli.py" = ["E402"]
"tests/test_ha_api.py" = ["E402"]
"scripts/*" = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
